utility function to fetch activity data
